架構拆分 (bot, app, domain) (#40)
