--- src/main/kotlin/tw/waterballsa/alpha/wsabot/gamification/commands/Gamification.kt
@@ -1,11 +1,0 @@
-package tw.waterballsa.alpha.wsabot.gamification.commands
-
-import me.jakejmattson.discordkt.commands.commands
-
-fun gamification() = commands("gamification") {
-    slash("gamification", "A 'gamification' command.") {
-        execute {
-            respond("Start gamification!")
-        }
-    }
-}
