--- src/main/resources/bot.properties
@@ -1,3 +1,0 @@
-description=A WSA-ALPHA-Bot for Waterball-Software-Academy
-name=WSA-ALPHA-BotExampleBot
-url=https://github.com/Waterball-Software-Academy/WSA-ALPHA-Bot
