--- src/main/resources/secret.properties
@@ -1,1 +1,0 @@
-token=
